requirements.txt
aws-cdk-lib==2.171.1
constructs>=10.0.0,<11.0.0
coverage==7.6.9
flake8==3.9.0
pytest==6.2.5
